requirements.txt
uvicorn
fastapi
python-dotenv
groq
textblob